Title: 세 수의 합과 평균 구하기, Time: 8ms, Memory: 0MB, Status: Passed - Codetree

+++ 251210/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.cpp
@@ -6,7 +6,7 @@ int main() {
 
     cin >> a >> b >> c;
 
-    cout << a + b + c << endl << (double)(a + b + c) / 3;
+    cout << a + b + c << endl << (a + b + c) / 3;
 
     return 0;
 }
